# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OASIS Projet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserService = (typeof UserService)[keyof typeof UserService];

export const UserService = {
  dev: "dev",
  commercial: "commercial",
  client: "client",
  projet: "projet",
  direction: "direction",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  team: string;
  avatar?: string;
  isAdmin?: boolean;
  service?: UserService;
  canAccessAllServices?: boolean;
}

export type DashboardSummarySuggestedActionsItem = {
  label: string;
  description: string;
  link: string;
};

export interface Client {
  id: string;
  ref: string;
  name: string;
  sector: string;
  contractStatus: string;
  priority: string;
  projectCount: number;
  recentTickets: number;
  latestIntervention: string;
  assignedTeam: string;
  logo?: string;
}

export interface TechnicalPlan {
  objective: string;
  impactedComponents: string[];
  actionSequence: string[];
  prerequisites: string[];
  dependencies: string[];
  risks: string[];
  testsToRun: string[];
  complexityEstimate: string;
  effortEstimate: string;
}

export interface AnalysisResult {
  id: string;
  projectId: string;
  clientId: string;
  taskType: string;
  prompt: string;
  status: string;
  createdAt: string;
  executiveSummary: string;
  requestUnderstanding?: string;
  projectContext?: string;
  relatedHistory?: string;
  identifiedRisks?: string[];
  affectedAreas?: string[];
  impactedComponents?: string[];
  recommendedApproach?: string;
  suggestedSteps?: string[];
  dependencies?: string[];
  confidenceScore: number;
  overallStatus: string;
  tags?: string[];
  technicalPlan?: TechnicalPlan;
}

export interface Activity {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  user: string;
  clientName?: string;
  projectName?: string;
}

export interface DashboardSummary {
  welcomeMessage: string;
  totalClients: number;
  totalProjects: number;
  totalAnalyses: number;
  recentClients: Client[];
  recentProjects: Project[];
  recentAnalyses: AnalysisResult[];
  recentActivities: Activity[];
  suggestedActions: DashboardSummarySuggestedActionsItem[];
}

export interface Projet {
  id: string;
  clientId: string;
  ref: string;
  name: string;
  description?: string;
  status: string;
  prestationCount: number;
  lastUpdated: string;
}

export interface Prestation {
  id: string;
  projectId: string;
  clientId: string;
  ref: string;
  name: string;
  description: string;
  status: string;
  techStack: string[];
  health: string;
  recentIncidents: number;
  nextMilestone: string;
  lastUpdated: string;
  productionUrl: string;
  saveUrls?: string[];
}

export interface AnalysisRequest {
  prestationId: string;
  projectId: string;
  taskType: string;
  prompt: string;
  documentName?: string;
}

export interface Source {
  id: string;
  analysisId: string;
  type: string;
  category: string;
  title: string;
  date: string;
  origin: string;
  excerpt: string;
  relevanceScore: number;
}

export interface DiffFile {
  path: string;
  language: string;
  additions: number;
  deletions: number;
  before: string;
  after: string;
  riskLabel: string;
}

export interface DiffResult {
  analysisId: string;
  files: DiffFile[];
  summary: string;
  totalAdditions: number;
  totalDeletions: number;
}

export interface Preferences {
  theme: string;
  language: string;
  responseDetail: string;
  notifications: boolean;
  autoSave: boolean;
}

export type AdminProfileRecentLoginsItem = {
  date: string;
  ip: string;
  device: string;
};

export type AdminProfileSessionSettings = {
  timeout: number;
  maxSessions: number;
};

export interface AdminProfile {
  user: User;
  recentLogins: AdminProfileRecentLoginsItem[];
  activityHistory: Activity[];
  sessionSettings: AdminProfileSessionSettings;
}

export interface FeedbackRequest {
  action: string;
  comment?: string;
  relevanceRating?: number;
  flagIssue?: boolean;
}

export type LoginBody = {
  email: string;
};

export type GetClientsParams = {
  search?: string;
  sector?: string;
  status?: string;
};

export type GetAnalysisHistoryParams = {
  clientId?: string;
  projectId?: string;
  taskType?: string;
  status?: string;
};

export type SubmitFeedback200 = {
  success: boolean;
};