# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OASIS Projet API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: dashboard
    description: Dashboard
  - name: clients
    description: Client management
  - name: projects
    description: Project management
  - name: analysis
    description: AI Analysis
  - name: sources
    description: Analysis sources
  - name: diff
    description: Code diff
  - name: activity
    description: Activity feed
  - name: preferences
    description: User preferences
  - name: admin
    description: Admin profile

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Simulate SSO login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required: [email]
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard summary
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"

  /clients/{id}/projects:
    get:
      operationId: getClientProjects
      tags: [projects]
      summary: Get projects for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Projet"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projet"

  /projects/{id}/prestations:
    get:
      operationId: getProjectPrestations
      tags: [projects]
      summary: Get prestations for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Prestation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prestation"

  /prestations/{id}:
    get:
      operationId: getPrestation
      tags: [projects]
      summary: Get prestation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Prestation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prestation"

  /analysis/run:
    post:
      operationId: runAnalysis
      tags: [analysis]
      summary: Run AI analysis
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"

  /analysis/history:
    get:
      operationId: getAnalysisHistory
      tags: [analysis]
      summary: Get analysis history
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
        - name: taskType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Analysis history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisResult"

  /analysis/{id}:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get single analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"

  /sources/{analysisId}:
    get:
      operationId: getSources
      tags: [sources]
      summary: Get sources for analysis
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"

  /diff/{analysisId}:
    get:
      operationId: getDiff
      tags: [diff]
      summary: Get diff for analysis
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diff data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiffResult"

  /activity:
    get:
      operationId: getActivity
      tags: [activity]
      summary: Get activity feed
      responses:
        "200":
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get user preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
    put:
      operationId: updatePreferences
      tags: [preferences]
      summary: Update preferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"

  /admin/profile:
    get:
      operationId: getAdminProfile
      tags: [admin]
      summary: Get admin profile
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"

  /analysis/{id}/feedback:
    post:
      operationId: submitFeedback
      tags: [analysis]
      summary: Submit feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        team:
          type: string
        avatar:
          type: string
        isAdmin:
          type: boolean
        service:
          type: string
          enum: [dev, commercial, client, projet, direction]
        canAccessAllServices:
          type: boolean
      required: [id, name, email, role, team]

    DashboardSummary:
      type: object
      properties:
        welcomeMessage:
          type: string
        totalClients:
          type: integer
        totalProjects:
          type: integer
        totalAnalyses:
          type: integer
        recentClients:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        recentProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        recentAnalyses:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisResult"
        recentActivities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        suggestedActions:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              description:
                type: string
              link:
                type: string
            required: [label, description, link]
      required: [welcomeMessage, totalClients, totalProjects, totalAnalyses, recentClients, recentProjects, recentAnalyses, recentActivities, suggestedActions]

    Client:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
        name:
          type: string
        sector:
          type: string
        contractStatus:
          type: string
        priority:
          type: string
        projectCount:
          type: integer
        recentTickets:
          type: integer
        latestIntervention:
          type: string
        assignedTeam:
          type: string
        logo:
          type: string
      required: [id, ref, name, sector, contractStatus, priority, projectCount, recentTickets, latestIntervention, assignedTeam]

    Projet:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        ref:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
        prestationCount:
          type: integer
        lastUpdated:
          type: string
      required: [id, clientId, ref, name, status, prestationCount, lastUpdated]

    Prestation:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        clientId:
          type: string
        ref:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
        techStack:
          type: array
          items:
            type: string
        health:
          type: string
        recentIncidents:
          type: integer
        nextMilestone:
          type: string
        lastUpdated:
          type: string
        productionUrl:
          type: string
        saveUrls:
          type: array
          items:
            type: string
      required: [id, projectId, clientId, ref, name, status, techStack, health, recentIncidents, nextMilestone, description, lastUpdated, productionUrl]

    AnalysisRequest:
      type: object
      properties:
        prestationId:
          type: string
        projectId:
          type: string
        taskType:
          type: string
        prompt:
          type: string
        documentName:
          type: string
      required: [prestationId, projectId, taskType, prompt]

    AnalysisResult:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        clientId:
          type: string
        taskType:
          type: string
        prompt:
          type: string
        status:
          type: string
        createdAt:
          type: string
        executiveSummary:
          type: string
        requestUnderstanding:
          type: string
        projectContext:
          type: string
        relatedHistory:
          type: string
        identifiedRisks:
          type: array
          items:
            type: string
        affectedAreas:
          type: array
          items:
            type: string
        impactedComponents:
          type: array
          items:
            type: string
        recommendedApproach:
          type: string
        suggestedSteps:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            type: string
        confidenceScore:
          type: number
        overallStatus:
          type: string
        tags:
          type: array
          items:
            type: string
        technicalPlan:
          $ref: "#/components/schemas/TechnicalPlan"
      required: [id, projectId, clientId, taskType, prompt, status, createdAt, executiveSummary, confidenceScore, overallStatus]

    TechnicalPlan:
      type: object
      properties:
        objective:
          type: string
        impactedComponents:
          type: array
          items:
            type: string
        actionSequence:
          type: array
          items:
            type: string
        prerequisites:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        testsToRun:
          type: array
          items:
            type: string
        complexityEstimate:
          type: string
        effortEstimate:
          type: string
      required: [objective, impactedComponents, actionSequence, prerequisites, dependencies, risks, testsToRun, complexityEstimate, effortEstimate]

    Source:
      type: object
      properties:
        id:
          type: string
        analysisId:
          type: string
        type:
          type: string
        category:
          type: string
        title:
          type: string
        date:
          type: string
        origin:
          type: string
        excerpt:
          type: string
        relevanceScore:
          type: number
      required: [id, analysisId, type, category, title, date, origin, excerpt, relevanceScore]

    DiffResult:
      type: object
      properties:
        analysisId:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/DiffFile"
        summary:
          type: string
        totalAdditions:
          type: integer
        totalDeletions:
          type: integer
      required: [analysisId, files, summary, totalAdditions, totalDeletions]

    DiffFile:
      type: object
      properties:
        path:
          type: string
        language:
          type: string
        additions:
          type: integer
        deletions:
          type: integer
        before:
          type: string
        after:
          type: string
        riskLabel:
          type: string
      required: [path, language, additions, deletions, before, after, riskLabel]

    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
        user:
          type: string
        clientName:
          type: string
        projectName:
          type: string
      required: [id, type, description, timestamp, user]

    Preferences:
      type: object
      properties:
        theme:
          type: string
        language:
          type: string
        responseDetail:
          type: string
        notifications:
          type: boolean
        autoSave:
          type: boolean
      required: [theme, language, responseDetail, notifications, autoSave]

    AdminProfile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        recentLogins:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              ip:
                type: string
              device:
                type: string
            required: [date, ip, device]
        activityHistory:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        sessionSettings:
          type: object
          properties:
            timeout:
              type: integer
            maxSessions:
              type: integer
          required: [timeout, maxSessions]
      required: [user, recentLogins, activityHistory, sessionSettings]

    FeedbackRequest:
      type: object
      properties:
        action:
          type: string
        comment:
          type: string
        relevanceRating:
          type: integer
        flagIssue:
          type: boolean
      required: [action]