# lib/api-zod/src/generated/types/analysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OASIS Projet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisRequest {
  prestationId: string;
  projectId: string;
  taskType: string;
  prompt: string;
  documentName?: string;
}