# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OASIS Projet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: string;
  ref: string;
  name: string;
  sector: string;
  contractStatus: string;
  priority: string;
  projectCount: number;
  recentTickets: number;
  latestIntervention: string;
  assignedTeam: string;
  logo?: string;
}