# lib/api-zod/src/generated/types/prestation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OASIS Projet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Prestation {
  id: string;
  projectId: string;
  clientId: string;
  ref: string;
  name: string;
  description: string;
  status: string;
  techStack: string[];
  health: string;
  recentIncidents: number;
  nextMilestone: string;
  lastUpdated: string;
  productionUrl: string;
  saveUrls?: string[];
}