# lib/api-zod/src/generated/types/projet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OASIS Projet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Projet {
  id: string;
  clientId: string;
  ref: string;
  name: string;
  description?: string;
  status: string;
  prestationCount: number;
  lastUpdated: string;
}